project_3/cloud_app/requirements.txt
streamlit==1.11.1
scikit-learn==1.0.2
pandas==1.4.2
numpy==1.21.5
pillow==9.0.1
nltk==3.7
